requirements.txt
customtkinter==5.1.2
pandas==1.5.2
pandastable==0.13.0
Pillow==10.0.1
